chore: Move file structure and update paths (#14)
