+++ src/pages/listPage/Chart/components/IdolCard.jsx
