+++ src/pages/listPage/Chart/components/IdolCard.module.css
